subindo boletim, + devidos ajustes de tipagem/rodando tudo ivanzão

+++ src/main/java/br/grupointegrado/educacional/dto/NotaRequestDTO.java
@@ -1,6 +1,7 @@
 package br.grupointegrado.educacional.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record NotaRequestDTO(Integer nota, LocalDate data_lancamento, Integer matricula_id, Integer disciplina_id) {
+public record NotaRequestDTO(BigDecimal nota, LocalDate data_lancamento, Integer matricula_id, Integer disciplina_id) {
 }
